add the binary file writer logic

+++ common/src/main/scala/org/bosch/common/domain/Measurement.scala
@@ -10,9 +10,9 @@ import scodec.codecs._
  * @param signalId id of the signal that produced the measurement
  * @param value it's value
  */
-final case class Measurement(timeSec: Int, timeUsec: Int, signalId: Int, value: Double)
+final case class Measurement(timeSec: Long, timeUsec: Long, signalId: Int, value: Double)
 
 object Measurement {
   val ReservedSize = 4
-  implicit val codec: Codec[Measurement] = (uint16 :: uint16 :: uint16 :: ignore(ReservedSize) :: double).as[Measurement]
+  implicit val codec: Codec[Measurement] = (uint32 :: uint32 :: uint16 :: ignore(ReservedSize) :: double).as[Measurement]
 }
